pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-copycat-src2id"
version = "1.1.2"
description = "Source to ID - Identify package coordinates and repositories from source code using multiple strategies"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "AGPL-3.0"}
authors = [
    {name = "Oscar Valenzuela B.", email = "oscar.valenzuela.b@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Version Control",
]

dependencies = [
    "aiohttp>=3.9.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "tabulate>=0.9.0",
]

[project.urls]
"Homepage" = "https://github.com/oscarvalenzuelab/semantic-copycat-swh"
"Bug Tracker" = "https://github.com/oscarvalenzuelab/semantic-copycat-swh/issues"
"Source Code" = "https://github.com/oscarvalenzuelab/semantic-copycat-swh"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]
swhid = [
    "miniswhid>=0.1.0",
    "swh.model>=6.0.0",
    "swh.web>=1.0.0",
]
enhanced = [
    "oslili>=0.1.0",
]


[project.scripts]
src2id = "src2id.cli.main:main"
src2id-validate = "src2id.cli.validate:validate_swhid"

[tool.setuptools.packages.find]
where = ["."]
include = ["src2id*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "C90", "I", "N", "UP", "B", "A", "C4", "PT", "SIM", "RUF"]
ignore = ["E501", "B008"]
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src2id --cov-report=term-missing"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src2id"]
omit = ["tests/*", "*/migrations/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]